src/requirements-data-gen.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements-data-gen.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.8
    # via
    #   anyio
    #   httpx
jiter==0.5.0
    # via openai
openai==1.42.0
    # via -r requirements-data-gen.in
pydantic==2.8.2
    # via
    #   -r requirements-data-gen.in
    #   openai
pydantic-core==2.20.1
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements-data-gen.in
pyyaml==6.0.2
    # via -r requirements-data-gen.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tenacity==9.0.0
    # via -r requirements-data-gen.in
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
